+++ src/com/intellij/idea/plugin/hybris/project/view/JunkProjectViewNode.kt
@@ -16,7 +16,7 @@
  * You should have received a copy of the GNU Lesser General Public License
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
-package com.intellij.idea.plugin.hybris.view
+package com.intellij.idea.plugin.hybris.project.view
 
 import com.intellij.ide.projectView.PresentationData
 import com.intellij.ide.projectView.ProjectViewNode