+++ src/com/intellij/idea/plugin/hybris/system/type/meta/TSMetaModelAccess.kt
@@ -1,6 +1,6 @@
 /*
- * This file is part of "SAP Commerce Developers Toolset" plugin for Intellij IDEA.
- * Copyright (C) 2019-2023 EPAM Systems <hybrisideaplugin@epam.com> and contributors
+ * This file is part of "SAP Commerce Developers Toolset" plugin for IntelliJ IDEA.
+ * Copyright (C) 2019-2024 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -48,5 +48,6 @@ interface TSMetaModelAccess {
     fun <T : TSGlobalMetaClassifier<*>> getAll(metaType: TSMetaType): Collection<T>
     fun getAll(): Collection<TSGlobalMetaClassifier<*>>
     fun getNextAvailableTypeCode(): Int
+    fun getReservedTypeCodes(): Map<Int, String>
     fun getAllOf(vararg metaTypes: TSMetaType): Collection<TSGlobalMetaClassifier<*>>
 }